Ignore artifactErrors (PHNX-10197) (#23)
Motivation
---
We want to ignore artifactErrors because renovate was complaining about not being able to find some of our private package, and so blocking merges. However, we already knew that a lot of our packages renovate won't be able to find because we have to manually add each one to a white list. And we couldn't find any other readily available documentation on what other kind of artifact errors it could find

Modifications
---
Ignore artifactErrors

Results
---
Stop blocking merges

https://centeredge.atlassian.net/browse/PHNX-10197